Add refine selection command
Allows refining a selection by selecting certain substrings. Very useful
in conjunction with multi-align.

The behaviour is configurable in the user's config.cson, and based on
current editor scope, allowing (and indeed requiring) language-specific
settings.

For example, in Haskell source, the following substrings inside the
current selections will be seleted:  = | :: -> <- , [