Clean up command registry of custom scripts
Previously, the commands were registered as a side-effect of each module
being included. This is now changed so that each module exports its
command list, and the script loading these files does the command
registration.

This modularization is a first step towards making my addons a proper
module, particularly one with a testsuite.